+++ NEWS.md
@@ -1,3 +1,5 @@
+# contactsurveys (development version)
+
 # contactsurveys 0.1.0
 
 This is a new package split off from the `{socialmixr}` package. It handles all interaction with the Zenodo repository containing social contact surveys, including listing available data sets, downloading data sets, and obtaining citations.